Deal properly with neighbor list before and after kernel invocation.
There are three(!) neighbor list representations.
1) LAMMPS
for ii in 0..inum:
  i = ilist[ii]               #< this mapping is identity for our inputs
  for jj in 0..numneigh[i]:
    j = firstneigh[i][jj]     #< where firstneigh[i] is a pointer into a set of pages
2) UNPICKLE
for e in 0..input->nedge:
  i = input->edge[(e*2)  ]
  j = input->edge[(e*2)+1]
3) GPU
for i in 0..N:
  for jj in 0..numneigh[i]:
    j = neighidx[(i*NSLOT)+jj]

I don't know if it is necessary to load/unload per-neighbor data (neighidx, shear and touch) before/after each kernel invocation.
It depends on if these values are required (and possibly mutated) between invocations.
This is a problem that still needs to be solved.

OTHER CHANGES
1) Make copies of all data mutated by kernel invocation. This ensures that all kernel invocations receive the same input.
2) Use input parameter file to initialize hertz constants.